tests/test_models/test_file_storage.py
#!/usr/bin/python3
"""Unittest for FileStorage class"""
import unittest
import os
from models.base_model import BaseModel
from models.engine.file_storage import FileStorage


class TestFileStorage(unittest.TestCase):
    """Test cases for FileStorage class"""

    def setUp(self):
        """Sets the class/obj"""
        self.storage = FileStorage()

    def tearDown(self):
        """Removes JSON file after testing the other methods"""
        if os.path.exists(FileStorage._FileStorage__file_path):
            os.remove(FileStorage._FileStorage__file_path)

    def test_all(self):
        """Test case for 'all' method"""
        self.assertEqual({}, self.storage.all())

    def test_new(self):
        """Test case for 'new' method"""
        model = BaseModel()
        self.storage.new(model)
        key = model.__class__.__name__ + "." + model.id
        self.assertIn(key, self.storage.all())

    def test_save(self):
        """Test case for 'save' method"""
        model = BaseModel()
        self.storage.new(model)
        self.storage.save()
        self.assertTrue(os.path.exists(FileStorage._FileStorage__file_path))

    def test_reload(self):
        """Test case for 'reload' method"""
        model = BaseModel()
        self.storage.new(model)
        self.storage.save()
        new_storage = FileStorage()
        new_storage.reload()
        key = model.__class__.__name__ + "." + model.id
        self.assertIn(key, new_storage.all())


if __name__ == "__main__":
    unittest.main()
